Use uint32_t instead of size_t to try make it more portable

+++ bfd/polymorphism.h
@@ -1,9 +1,9 @@
-#include <stddef.h>    // size_t
+#include <stdint.h>    // uint32_t
 extern char **Duplicate_Argv_Plus_Extra(int const argc, char **argv, char const *const extra);
 extern char unsigned const (*Polymorphism_Get_128_Random(void))[16u];
 extern char const *Polymorphism_Get_Name_Extra_Object_File(void);
 extern int Polymorphism_Process_Symbol_1st_Run(char const*);
-extern int Polymorphism_Process_Symbol_2nd_Run(char const*, size_t);
-extern size_t Polymorphism_Finalise_Symbols_And_Create_Map(void**);
+extern int Polymorphism_Process_Symbol_2nd_Run(char const*, uint32_t);
+extern uint32_t Polymorphism_Finalise_Symbols_And_Create_Map(void**);
 extern char const *Polymorphism_Create_Extra_Object_File(void);
